requirements.txt
streamlit
tensorflow-macos
tensorflow-metal
keras
numpy
nltk